Created a new factory for repositories
Created a new factory for repositories, so that there is one
call to be able to get configured repository. This will allow
us to start using other repositories potentially.
