+++ tests/implementations/fedora3/FedoraApi/ApiUploadTest.php
@@ -1,16 +1,13 @@
 <?php
+
+require_once 'RepositoryFactory.php';
 require_once 'tests/TestHelpers.php';
 
 class UploadTest extends PHPUnit_Framework_TestCase {
 
   protected function setUp() {
-    $connection = new RepositoryConnection(FEDORAURL, FEDORAUSER, FEDORAPASS);
-    $this->api = new FedoraApi($connection);
-
-    $connection = new RepositoryConnection(FEDORAURL, FEDORAUSER, FEDORAPASS);
-    $this->api = new FedoraApi($connection);
-    $cache = new SimpleCache();
-    $this->repository = new FedoraRepository($this->api, $cache);
+    $this->repository = RepositoryFactory::getRepository('fedora3', new RepositoryConfig(FEDORAURL, FEDORAUSER, FEDORAPASS));
+    $this->api = $this->repository->api;
   }
 
   public function testUploadString() {
