+++ tests/implementations/fedora3/FedoraRelationships/FedoraRelationshipsTest.php
@@ -1,5 +1,7 @@
 <?php
-require_once "implementations/fedora3/FedoraRelationships.php";
+require_once 'RepositoryFactory.php';
+require_once 'tests/TestHelpers.php';
+
 /**
  * @todo pull more tests out of tjhe microservices version of these functions
  *  to make sure we handle more cases.
@@ -8,13 +10,12 @@ require_once "implementations/fedora3/FedoraRelationships.php";
  *  domdocument cannonicalization function that doesn't work properly on cent
  */
 class FedoraRelationshipsTest extends PHPUnit_Framework_TestCase {
+  function setUp() {
+    $this->repository = RepositoryFactory::getRepository('fedora3', new RepositoryConfig(FEDORAURL, FEDORAUSER, FEDORAPASS));
+  }
 
   function testRelationshipDescription() {
-    $connection = new RepositoryConnection(FEDORAURL, FEDORAUSER, FEDORAPASS);
-    $this->api = new FedoraApi($connection);
-    $cache = new SimpleCache();
-    $repository = new FedoraRepository($this->api, $cache);
-    $object = $repository->constructObject("test:test");
+    $object = $this->repository->constructObject("test:test");
     $rel = $object->relationships;
 
     $rel->registerNamespace('fuckyah', 'http://crazycool.com#');
@@ -39,11 +40,7 @@ class FedoraRelationshipsTest extends PHPUnit_Framework_TestCase {
   </description>
 </RDF>
 XML;
-    $connection = new RepositoryConnection(FEDORAURL, FEDORAUSER, FEDORAPASS);
-    $this->api = new FedoraApi($connection);
-    $cache = new SimpleCache();
-    $repository = new FedoraRepository($this->api, $cache);
-    $object = $repository->constructObject("test:test");
+    $object = $this->repository->constructObject("test:test");
     $datastream = $object->constructDatastream('RELS-EXT', 'M');
     $datastream->content = $content;
     $object->ingestDatastream($datastream);
@@ -64,11 +61,7 @@ XML;
   </rdf:Description>
 </rdf:RDF>
 XML;
-    $connection = new RepositoryConnection(FEDORAURL, FEDORAUSER, FEDORAPASS);
-    $this->api = new FedoraApi($connection);
-    $cache = new SimpleCache();
-    $repository = new FedoraRepository($this->api, $cache);
-    $object = $repository->constructObject("islandora:479");
+    $object = $this->repository->constructObject("islandora:479");
     $datastream = $object->constructDatastream('RELS-EXT', 'M');
     $datastream->content = $content;
     $object->ingestDatastream($datastream);
@@ -89,12 +82,7 @@ $expected = <<<XML
 </rdf:RDF>
 
 XML;
-
-    $connection = new RepositoryConnection(FEDORAURL, FEDORAUSER, FEDORAPASS);
-    $this->api = new FedoraApi($connection);
-    $cache = new SimpleCache();
-    $repository = new FedoraRepository($this->api, $cache);
-    $object = $repository->constructObject("test:test");
+    $object = $this->repository->constructObject("test:test");
     $rel = $object->relationships;
 
     $rel->registerNamespace('fuckyah', 'http://crazycool.com#');
