+++ tests/implementations/fedora3/Object/ObjectTest.php
@@ -1,19 +1,13 @@
 <?php
 
-require_once 'implementations/fedora3/FedoraApi.php';
-require_once 'implementations/fedora3/FedoraApiSerializer.php';
-require_once 'implementations/fedora3/Object.php';
-require_once 'implementations/fedora3/Repository.php';
-require_once 'Cache.php';
+require_once 'RepositoryFactory.php';
 require_once 'tests/TestHelpers.php';
 
 class ObjectTest extends PHPUnit_Framework_TestCase {
 
   protected function setUp() {
-    $connection = new RepositoryConnection(FEDORAURL, FEDORAUSER, FEDORAPASS);
-    $this->api = new FedoraApi($connection);
-    $cache = new SimpleCache();
-    $repository = new FedoraRepository($this->api, $cache);
+    $this->repository = RepositoryFactory::getRepository('fedora3', new RepositoryConfig(FEDORAURL, FEDORAUSER, FEDORAPASS));
+    $this->api = $this->repository->api;
 
     // create an object
     $string1 = FedoraTestHelpers::randomString(10);
