+++ tests/implementations/fedora3/RepositoryTest.php
@@ -1,18 +1,13 @@
 <?php
 
-require_once 'implementations/fedora3/FedoraApiSerializer.php';
-require_once 'implementations/fedora3/Object.php';
-require_once 'implementations/fedora3/Repository.php';
-require_once 'implementations/fedora3/RepositoryConnection.php';
-require_once 'implementations/fedora3/FedoraApi.php';
+require_once 'RepositoryFactory.php';
+require_once 'tests/TestHelpers.php';
 require_once 'tests/implementations/RepositoryTestBase.php';
 
 class Fedora3RepositoryTest extends RepositoryTestBase {
 
   protected function setUp() {
-    $connection = new RepositoryConnection(FEDORAURL, FEDORAUSER, FEDORAPASS);
-    $this->api = new FedoraApi($connection);
-    $cache = new SimpleCache();
-    $this->repository = new FedoraRepository($this->api, $cache);
+    $this->repository = RepositoryFactory::getRepository('fedora3', new RepositoryConfig(FEDORAURL, FEDORAUSER, FEDORAPASS));
+    $this->api = $this->repository->api;
   }
 }
